src: re-structure repository to delivery
The project had been developed in a private repository so far, so it was
structured without thinking about other files in the main folder, like
CODE_OF_CONDUCT and CONTRIBUTING.

To make the files easier to understand, create a new folder 'BotBlock'
and move all our work to there.
