+++ BotBlock/manifest.json
