+++ BotBlock/tools.js
